refactor(harness): extract HarnessId into leaf module PureClaw.Harness.Id
Move the HarnessId newtype, parseHarnessId, harnessIdToText, and its
ToJSON/FromJSON instances from Registry.hs into a new dependency-free leaf
module (PureClaw.Harness.Id). Registry re-exports these names unchanged so
all ~30 existing importers are untouched. Session.Kind now imports HarnessId
from the leaf, breaking the latent Registry <-> Session.Kind import cycle that
would otherwise arise when Registry gains a HarnessFlavour field (from
Session.Kind) in the next task.

Build: 141/141 modules, 0 warnings under -Wall -Werror.
Tests: 2787 examples, 0 failures.
